+++ Assets/Scripts/Combat_Refactored/Interfaces/IHaveHealth.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IHaveHealth
 {
-    int Health { get; set; }
+    float Health { get; set; }
 }
